+++ pgvector/django/__init__.py
@@ -1,6 +1,7 @@
 from .extensions import VectorExtension
 from .functions import L2Distance, MaxInnerProduct, CosineDistance, L1Distance
+from .halfvec import HalfvecField
 from .indexes import IvfflatIndex, HnswIndex
 from .vector import VectorField
 
-__all__ = ['VectorExtension', 'VectorField', 'IvfflatIndex', 'HnswIndex', 'L2Distance', 'MaxInnerProduct', 'CosineDistance', 'L1Distance']
+__all__ = ['VectorExtension', 'VectorField', 'HalfvecField', 'IvfflatIndex', 'HnswIndex', 'L2Distance', 'MaxInnerProduct', 'CosineDistance', 'L1Distance']