+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.6 (unreleased)
 
+- Added support for `halfvec` type to Django
 - Added support for `halfvec` and `sparsevec` types to Psycopg 3
 - Added support for `halfvec` and `sparsevec` types to Psycopg 2
 - Added support for `halfvec` and `sparsevec` types to asyncpg